Fixed level problem (-1), and bad background path

+++ src/hu/itsh/gyakorlat/szotar/io/excel/ExcelBase.java
@@ -63,7 +63,7 @@ public class ExcelBase {
 						rowStringBuilder.append(new SimpleDateFormat(SharedConstants.FORMAT_YYMMDD).format(date));
 						rowStringBuilder.append(separator);
 					} else {
-						rowStringBuilder.append(nextCell.getNumericCellValue());
+						rowStringBuilder.append((int)nextCell.getNumericCellValue());
 						rowStringBuilder.append(separator);
 					}
